+++ src/main/java/com/iexec/core/Application.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,10 +19,11 @@ package com.iexec.core;
 import io.changock.runner.spring.v5.config.EnableChangock;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @EnableChangock
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class Application {
 
     public static void main(String[] args) {