+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Add and use a non-root user in the dockerfile. (#627)
 - Replace single thread executor with synchronized keyword. (#633)
 - Move contribution status checks from `iexec-commons-poco`. (#636)
+- Use `BlockchainAdapterService` from `iexec-blockchain-adapter-api-library`. (#641)
 
 ### Bug Fixes